requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pandas==2.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
playwright==1.41.2
chardet==5.2.0
python-multipart==0.0.9
typing-extensions==4.9.0
numpy==1.23.5
